# Improved scart performance