Current scart implementation prevent dumping a large amount of stations and channels because the time spent in sorting the records is way, way larger than the IO time needed to read/write the data.

This patch fix this behaviour so that scart is no longer a bottleneck. The patch make use of bisect module to keep the list of stream/records constantly sorted.


Here is the scart profiler output when trying to dump 20 minutes of data from ~700 stations (3 channels each):

```
python -m cProfile -o prof /usr/local/bin/scart -dsE -t '2016/07/20 03:24:00~2016/07/20 03:44:00' -n '8X,BW,CH,CR,CZ,FR,G,GE,GR,GU,HS,HU,IU,IV,MN,NI,OE,OX,PL,RA,RD,RF,RO,SI,SK,SL,ST,SX,TH,XG,Z3' -c '(B|H)H(Z|N|E)' --files 2500  /alparray/ > sorted.mseed

>>> import pstats
>>> p = pstats.Stats('prof')
>>> p.sort_stats('tottime').print_stats(20)
Thu Nov  9 12:14:06 2017    prof

         2057424514 function calls (2057418376 primitive calls) in 2563.958 seconds

   Ordered by: internal time
   List reduced from 370 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
   500804 1115.558    0.002 2168.306    0.004 /usr/local/bin/scart:282(nextSort)
648278955  415.844    0.000  907.467    0.000 /usr/local/bin/scart:249(__cmp__)
648278955  260.655    0.000  260.655    0.000 {_Core.TimeSpan___gt__}
648278955  230.204    0.000  490.858    0.000 /usr/local/lib/python/seiscomp3/Core.py:580(__gt__)
  5460296  223.149    0.000  223.149    0.000 {_IO.RecordInput_next}
     4514  141.942    0.031  141.942    0.031 {posix.listdir}
     2269   41.561    0.018  216.274    0.095 /usr/local/bin/scart:128(findIndex)
     4425   17.553    0.004   17.553    0.004 {_IO.FileRecordStream_setSource}
        1   12.330   12.330 2563.958 2563.958 /usr/local/bin/scart:15(<module>)
   504803    8.836    0.000    8.836    0.000 {range}
  5963191    8.751    0.000    8.751    0.000 {_Core.Record_endTime}
   500803    5.476    0.000  135.632    0.000 /usr/local/bin/scart:226(next)
   500804    5.239    0.000    5.239    0.000 {_Core.Time_GMT}
  5462440    4.803    0.000    4.803    0.000 {_Core.Record_startTime}
  4961701    4.505    0.000  129.251    0.000 /usr/local/lib/python/seiscomp3/IO.py:479(__iter__)
  5460284    4.136    0.000    4.136    0.000 {_Core.Record_samplingFrequency}
   500803    4.074    0.000  126.973    0.000 /usr/local/bin/scart:157(readRecord)
  5460296    4.021    0.000  227.170    0.000 /usr/local/lib/python/seiscomp3/IO.py:478(next)
  5460284    3.945    0.000    7.660    0.000 /usr/local/lib/python/seiscomp3/Core.py:581(__ge__)
  5460284    3.721    0.000    7.857    0.000 /usr/local/lib/python/seiscomp3/Core.py:879(samplingFrequency)
```

You can clearly see there is more time spent in sorting the records than actual I/O

Those are the performance of the patched scart:

```
python -m cProfile -o prof ./fast-scart.py -dsE -t '2016/07/20 03:24:00~2016/07/20 03:44:00' -n '8X,BW,CH,CR,CZ,FR,G,GE,GR,GU,HS,HU,IU,IV,MN,NI,OE,OX,PL,RA,RD,RF,RO,SI,SK,SL,ST,SX,TH,XG,Z3' -c '(B|H)H(Z|N|E)' --files 2500  /alparray/ > sorted.mseed


>>> import pstats
>>> p = pstats.Stats('prof')
>>> p.sort_stats('tottime').print_stats(20)
Thu Nov  9 17:34:31 2017    prof

         132868555 function calls (132862417 primitive calls) in 463.104 seconds

   Ordered by: internal time
   List reduced from 370 to 20 due to restriction <20>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  5460296  167.121    0.000  167.121    0.000 {_IO.RecordInput_next}
     4514  112.666    0.025  112.666    0.025 {posix.listdir}
     2269   38.575    0.017  205.992    0.091 ./fast-scart2.py:128(findIndex)
     4425   16.563    0.004   16.563    0.004 {_IO.FileRecordStream_setSource}
   500803   12.401    0.000   26.223    0.000 {_bisect.insort}
        1   11.006   11.006  463.104  463.104 ./fast-scart2.py:15(<module>)
  5963191    7.780    0.000    7.780    0.000 {_Core.Record_endTime}
  5555272    6.129    0.000   13.822    0.000 ./fast-scart2.py:249(__cmp__)
   500803    5.068    0.000   82.125    0.000 ./fast-scart2.py:226(next)
   500804    4.520    0.000    4.520    0.000 {_Core.Time_GMT}
  5462440    4.224    0.000    4.224    0.000 {_Core.Record_startTime}
  5460284    4.169    0.000    4.169    0.000 {_Core.Record_samplingFrequency}
  4961701    4.164    0.000  124.511    0.000 /usr/local/lib/python/seiscomp3/IO.py:479(__iter__)
  7848444    3.976    0.000    7.856    0.000 /usr/local/lib/python/seiscomp3/Core.py:578(__lt__)
  5460296    3.889    0.000  171.010    0.000 /usr/local/lib/python/seiscomp3/IO.py:478(next)
  7848444    3.880    0.000    3.880    0.000 {_Core.TimeSpan___lt__}
   500803    3.813    0.000   73.959    0.000 ./fast-scart2.py:157(readRecord)
  5460284    3.678    0.000    7.290    0.000 /usr/local/lib/python/seiscomp3/Core.py:581(__ge__)
  5460284    3.612    0.000    3.612    0.000 {_Core.TimeSpan___ge__}
  5460284    3.515    0.000    7.683    0.000 /usr/local/lib/python/seiscomp3/Core.py:879(samplingFrequency)

```

Total time decreased from 2563 seconds to 463 seconds and most of the time is spent in I/O as it should