@gempa-jabe 2017-11-10
Thanks a lot. The speed up looks great. Indeed, the current implementation assumed that sort (qsort) on an already sorted array should be fast. The bisect module probably touches less objects and involves less comparisons compared to a full sort. I think this could be even much faster when implemented in C++ but that is out of scope of this PR.